test: 쿠폰 차감 동시성 테스트
